% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_06_open_ended_coding_code_gpt_to_excel.R
\name{code_gpt_to_excel}
\alias{code_gpt_to_excel}
\title{\strong{Write \code{code_gpt()} results to Excel workbook}}
\usage{
code_gpt_to_excel(coded_tbl, path = "Data")
}
\arguments{
\item{coded_tbl}{A dataset returned by \code{code_gpt()} with coding results.}

\item{path}{Directory containing the coding workbook. Defaults to \code{"Data"}.}
}
\value{
The Excel workbook is updated in place.
}
\description{
This function inserts the coded responses (\code{Code(s)} column) from a \code{code_gpt()} result table into the corresponding Excel coding workbook created for that question.
}
\details{
\itemize{
\item The function locates the workbook named \verb{X Coding Workbook.xlsx}, matches coded responses by respondent ID, and writes the codes into the \code{Code(s)} of the coding sheet.
\item \code{import_coding()} function can be used to import and process the coding workbook.
}

Use this function immediately after \code{code_gpt()} in a pipeline to write the coded results directly to the corresponding Excel workbook.

You do not need to save it as a dataset, as the workbook is updated in place.

For example:

\preformatted{
code_gpt(data = survey_data, x = Q5, theme_list = theme_list) %>%
  code_gpt_to_excel()
}

This will update \verb{Q5 Coding Workbook} in place with the coded responses, which can then be imported using \code{import_coding()}.
}
\examples{
# Update "Q5 Coding Workbook" in place with the coded responses
code_gpt(data = survey_data, x = Q5, theme_list = theme_list) \%>\%
  code_gpt_to_excel()

}
