% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_03_misc_copy2clip.R
\name{copy2clip}
\alias{copy2clip}
\title{\strong{Copy to clipboard}}
\usage{
copy2clip(x, row.names = FALSE, col.names = TRUE, ...)
}
\arguments{
\item{x}{Object to copy.}

\item{row.names}{Include row names. Defaults to \code{FALSE}.}

\item{col.names}{Include column names. Defaults to \code{TRUE}.}
}
\description{
This function copies a data frame, matrix, or vector to the system clipboard as a tab-separated table (Windows only).
}
\examples{
# Copy a data frame to clipboard
copy2clip(data)

}
