% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_05_statistical_cor_heatmap.R
\name{cor_heatmap}
\alias{cor_heatmap}
\title{\strong{Plot a correlation heatmap}}
\usage{
cor_heatmap(cor_mat, exclude = NULL)
}
\arguments{
\item{cor_mat}{Correlation matrix or a data frame.}

\item{exclude}{(Optional) Variable(s) to exclude.}
}
\value{
A plotly heatmap object.
}
\description{
This function creates an interactive heatmap of a correlation matrix. Hover labels display variable pairs and their correlation values.
}
\details{
The function accepts the output of \code{cor_matrix()} and optionally allows excluding selected variables from the visualization.
}
\examples{
# Basic heatmap pipeline
cor_matrix(data = data) \%>\% cor_heatmap()

}
