% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_01_metadata_dat_sum_fn.R
\name{dat_sum_fn}
\alias{dat_sum_fn}
\title{\strong{Summarize dataset metadata}}
\usage{
dat_sum_fn(data)
}
\arguments{
\item{data}{A data frame to summarize.}
}
\value{
A tibble with a summary row for each variable.
}
\description{
This function produces a summary of each variable in a dataset, including variable name, class, label, number of non-missing values, percentage of non-missing values, and factor levels (if applicable).
}
\examples{
# Summarize survey_data
data_sum <- dat_sum_fn(survey_data)

}
