% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_03_statistical_one_hot_encode.R
\name{one_hot_encode}
\alias{one_hot_encode}
\title{\strong{One-hot encode categorical variables}}
\usage{
one_hot_encode(data, vars, keep_all = TRUE)
}
\arguments{
\item{data}{A data frame containing the variables to encode.}

\item{vars}{Variables to one-hot encode.}

\item{keep_all}{Logical; if \code{TRUE}, returns full dataset; if \code{FALSE}, returns only one-hot encoded variables.}
}
\value{
A data frame with one-hot encoded variables.
\itemize{
\item If \code{keep_all = TRUE}, the full dataset is returned with one-hot encoded variables replacing originals.
\item If \code{keep_all = FALSE}, only one-hot encoded variables are returned.
}
}
\description{
This function converts one or more categorical variables in a data frame into one-hot encoded variables.
}
\details{
\itemize{
\item Binary factors labeled \code{Unchecked}/\code{Checked} are encoded as a single 0–1 variable.
\item Multi-category factors are expanded into multiple one-hot encoded variables (one per level).
\item Original variable labels are preserved and extended to the generated one-hot encoded variables.
}
}
\examples{
# One-hot encode region and Q1_1 while keeping all original variables
data <- one_hot_encode(data = data, c(region, Q1_1))

# One-hot encode region and Q1_1, returning only the encoded columns
data_new <- one_hot_encode(data = data, c(region, Q1_1), keep_all = FALSE)

}
