% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_05_tabulation_tab_grid.R
\name{print.tab_grid_result}
\alias{print.tab_grid_result}
\title{Custom print method for tab_grid() results}
\usage{
\method{print}{tab_grid_result}(x, ...)
}
\description{
Custom print method for tab_grid() results
}
