% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_07_tabulation_tab_grid_t.R
\name{print.tab_grid_t_result}
\alias{print.tab_grid_t_result}
\title{Custom print method for tab_grid_t() results}
\usage{
\method{print}{tab_grid_t_result}(x, ...)
}
\description{
Custom print method for tab_grid_t() results
}
