% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_01_statistical_standardize.R
\name{standardize}
\alias{standardize}
\title{\strong{Standardize numeric and factor variables (z-scores)}}
\usage{
standardize(
  data,
  vars = everything(),
  flip = FALSE,
  keep_all = TRUE,
  outside_scale = c("Don't know", "Prefer not to say")
)
}
\arguments{
\item{data}{A data frame containing variables to standardize.}

\item{vars}{Variables to standardize. Supports tidyselect syntax (e.g., \code{starts_with()}, \code{c(var1, var2)}).}

\item{flip}{Logical; if \code{TRUE}, reverses the direction of valid response levels before standardization.}

\item{keep_all}{Logical; if \code{TRUE}, returns full dataset; if \code{FALSE}, returns only standardized variables.}

\item{outside_scale}{Character vector of response levels (typically DK/PNS) that lie outside of the scale and should be treated as neutral midpoint (\code{0.5}).}
}
\value{
A data frame with standardized variables.
\itemize{
\item If \code{keep_all = TRUE}, the full dataset is returned with standardized variables.
\item If \code{keep_all = FALSE}, only standardized variables are returned.
}
}
\description{
This function standardizes variables to have mean = 0 and standard deviation = 1.
}
\details{
\itemize{
\item Numeric variables are z-scored.
\item Factor variables are converted to numeric order and then standardized.
\item Levels in \code{outside_scale} are assigned to the midpoint of the valid range before transformation.
}
}
\examples{
# Standardize Q1 variable set with
standardize(data = data, vars = starts_with("Q1_"))

# Standardize Q1 and Q2 variable set with the scale order
standardize(data = data, vars = c(Q1, Q2), flip = TRUE)

}
