% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_tabulation_tab_all.R
\name{tab_all}
\alias{tab_all}
\title{\strong{Apply \code{tab()} to multiple variables}}
\usage{
tab_all(data, y = NULL, weight = NULL, exclude = NULL, ...)
}
\arguments{
\item{data}{A data frame containing the survey data.}

\item{y}{(Optional) A variable to cross-tabulate against all \code{x} variables.}

\item{exclude}{Variables to exclude from tabulation. Defaults to \code{NULL}. Always excludes \code{y} (if specified), \code{weight} (if specified), and variable literally named "weight".}

\item{...}{Additional arguments passed to \code{tab()}.}
}
\value{
A named list of frequency tables or crosstabs, one per variable.
}
\description{
This function maps \code{tab()} across all variables in a data frame.
Optionally, you can specify a \code{y} variable for crosstabs and/or exclude variables from being tabulated.
Additional arguments are passed to \code{tab()}.
}
\examples{
# Unweighted frequency tables of all questions
tab_all(data = survey_data)

# Weighted crosstabs of all questions by region, excluding Q99
tab_all(data = survey_data, y = region, weight = weight_var, exclude = "Q99")
}
