% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_tabulation_tab_mr.R
\name{print.tab_mr_result}
\alias{print.tab_mr_result}
\title{Custom print method for tab_mr() results}
\usage{
\method{print}{tab_mr_result}(x, ...)
}
\description{
Custom print method for tab_mr() results
}
